클라 방향추가, 데미지추가 =진서

+++ Client/Source/BigNanto/BigNantoGameInstance.h
@@ -93,7 +93,7 @@ public:
 	void PacketProcess(Packet& packet);
 
 	// 로그인 버퍼 생성 함수
-	int MakeLoginBuf(char * source, char cls, float Y, float Z, uint32 damage, char * name, int namelen);
+	int MakeLoginBuf(char * source, char cls, float Y, float Z, float damage, char * name, int namelen);
 	// 플레이어 입장
 	UFUNCTION(BlueprintCallable)
 	void EnterGame(FString ServerIP, int32 ServerPort, FString UserName, uint8 ClassType);
